Update summary pres
Updates to summary pres.   Now includes narration.  You can export this as a mp4 from PPT to get a video (I can't upload it as it is 47MB, bigger than the 25MB limit for github).  The audio needs some work (I need to write out the script, practice it, and re-record each slide until it is clean) but it will give you an idea.   Powerpoint also includes subtitle generation but it makes a lot of stupid errors and no editor is available to fix them.  I think I probably need to add the captions myself manually using a video editor, unfortunately.   I'll get the audio cleaned up first...