Add intro pres
this is an expanded version of the lightning talk, for use at TPAC joint meetings.
It introduces the basics of WoT, current work items, etc. to make sure everyone is oriented.
It also provides links to resources.  Unlike the lightning talk deck, this deck does not include slide timings or recorded narration.